Move gen_parser_function to codegen module
